Ensure the right scoping for variable sharing and checkpointing.
Currently, the scoping makes the graph very difficult to visualise, but
it nvertheless allows for correct variable sharing between the inference
cells and generator cells at various timesteps.

More time needs to be spent on understanding the interplay of
tf.name_scope, tf.variable_scope, and the default scoping mechanisms in
TF layers to make sure the graph is actually easily readable. This is
currently not a priority.