Extract GQNTFRecordDataset outside of input_fn.
In order to make the input_fn clearer, all the parsing functions that
required access to state variables have been consolidated as methods of
a GQNTFRecordDataset. This implements the tf.data.Dataset interface, so
it’s easy to use in the new input_fn which shuffles, buffers, and
repeats the dataset as requested.